+++ test/delighted_test.rb
@@ -7,11 +7,18 @@ class Delighted::ClientTest < Delighted::TestCase
   end
 
   def test_handles_rate_limited_response
-    response = Delighted::HTTPResponse.new(429, {}, Delighted::JSON.dump({ :status => 429, :message => "Too Many Requests" }
+    response = Delighted::HTTPResponse.new(429, { "Retry-After" => "10" }, Delighted::JSON.dump({ :status => 429, :message => "Too Many Requests" }
 ))
     mock_http_adapter.stubs(:request).returns(response)
 
-    assert_raises(Delighted::RateLimitedError) { Delighted.shared_client.get_json("/foo") }
+    assert_raises(Delighted::RateLimitedError) do
+      begin
+        Delighted.shared_client.get_json("/foo")
+      rescue Delighted::RateLimitedError => e
+        assert_equal 10, e.retry_after
+        raise
+      end
+    end
   end
 end